Major refactorings in the internal handling of IO registers (ActiveRegisters); removed the writeBit() method globally and refactored usages; moved many devices to being to use RegisterView and BooleanView instances to view subranges of registers with consistency; added EventBuffer and EventGen, which are eventually to subsume SimPrinter; removed Simulator.IORWatch

+++ src/avrora/sim/radio/Radio.java
@@ -67,7 +67,7 @@ public interface Radio {
         public final long originTime;
         public final long deliveryTime;
 
-        public int strength = ADC.VBG_LEVEL;
+        public float strength = ADC.VBG_LEVEL;
 
         public Transmission(byte data, long frequency, long originTime) {
             this.data = data;
